+++ lib/resty/auto-ssl/utils/start_sockproc.lua
@@ -10,7 +10,13 @@ local function start()
 
   ngx.log(ngx.NOTICE, "auto-ssl: starting sockproc")
 
-  local _, run_err = shell_blocking.capture_combined({ auto_ssl.lua_root .. "/bin/resty-auto-ssl/start_sockproc" }, { umask = "0022" })
+  local start_sockproc_path = auto_ssl:get_bin("start_sockproc")
+  if not start_sockproc_path then
+    ngx.log(ngx.ERR, "auto-ssl: unable to resolve the path to start_sockproc")
+    return
+  end
+
+  local _, run_err = shell_blocking.capture_combined({ start_sockproc_path }, { umask = "0022" })
   if run_err then
     ngx.log(ngx.ERR, "auto-ssl: failed to start sockproc: ", run_err)
   else